requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
frozendict==2.4.4
html5lib==1.1
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
MarkupSafe==2.1.5
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
plotly==5.23.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
retrying==1.3.4
scikit-learn==1.5.1
scipy==1.14.1
setuptools==73.0.1
six==1.16.0
soupsieve==2.6
tenacity==9.0.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.4
yfinance==0.2.43
zipp==3.20.0